requirements.txt
https://www.djangoproject.com/download/1.5c1/tarball/#egg=Django
Fabric==1.5.1
Markdown==2.2.1
PyGithub==1.10.0
Pygments==1.5
South==0.7.6
Werkzeug==0.8.3
django-appconf==0.5
django-compressor==1.2
django-debug-toolbar==0.9.4
django-disqus==0.4.1
django-extensions==1.0.2
django-grappelli==2.4.3
-e git+git://github.com/carljm/django-markitup.git#egg=markitup
django-model-utils==1.1.0
django-pagination==1.0.7
django-reversion==1.6.5
-e git+git://github.com/mbi/django-rosetta.git#egg=rosetta
-e git+git://github.com/zsiciarz/django-rosetta-grappelli.git#egg=rosetta_grappelli
# using a fork until a new PyPI release is made
-e git+git://github.com/hcarvalhoalves/django-taggit.git#egg=taggit
docutils==0.10
flake8==1.7.0
gunicorn==0.17.2
ipython==0.13.1
newrelic==1.6.0.13
paramiko==1.9.0
psycopg2==2.4.6
pycrypto==2.6
raven==3.1.0
